requirements.txt
amqp==2.6.0
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.3.3
autopep8==1.4.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.1
cryptography==2.9.2
decorator==4.4.2
distlib==0.3.1
dnspython==1.16.0
docxtpl==0.10.0
eventlet==0.25.2
filelock==3.0.12
Flask==1.1.2
Flask-APScheduler==1.11.0
greenlet==0.4.16
idna==2.10
imgkit==1.0.2
importlib-metadata==1.7.0
inflection==0.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
marshmallow==3.6.1
mccabe==0.6.1
mock==4.0.2
monotonic==1.5
mysqlclient==2.0.1
numpy==1.19.0
opencv-python==4.2.0.34
pandas==1.0.5
passlib==1.7.2
path==13.1.0
path.py==12.4.0
Pillow==7.2.0
prettytable==0.7.2
prompt-toolkit==1.0.14
psutil==5.7.0
pycodestyle==2.5.0
pycparser==2.20
pyfiglet==0.8.post1
Pygments==2.6.1
PyInquirer==1.0.3
PyJWT==1.7.1
pylint==2.4.4
PyQt5==5.15.0
PyQt5-sip==12.8.0
PySimpleGUI==4.24.0
python-dateutil==2.8.1
python-docx==0.8.10
python-telegram-bot==12.8
pytz==2020.1
pywin32==228
PyYAML==5.3.1
regex==2020.6.8
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.7
termcolor==1.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.47.0
typed-ast==1.4.0
tzlocal==2.1
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.25
waitress==1.4.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.11.2
zipp==3.1.0
